+++ src/lib/components/ui/fallbacks/display-trip-fallback.tsx
